TI thermal framework: improve thermal domain manipulations
This patch changes the thermal device data structures
to have a pointer to thermal domain. This way, once
an operation is required on a thermal device, there is
no need to iterate on thermal domains lists.

This patch also centralizes the operations to search and create
thermal domain data structures. Idea is to improve code readability,
having these operations done in one place only.

The patch does not change the code functionality so far.

Change-Id: I482d82feac854270c300273efbef0ef07cdec0b2
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>